Add files via upload
This Python script authenticates with the Google Sheets API. It then builds a connection to the API and prepares to update a specific Google Sheet file. Then after defining the range and sending the list of values, it will execute the update and print a response confirming that the data was stored in Google Sheets.